* Change the return from FileUtils::size() to an int so it can   distinguish between an error and a truly zero-length file

+++ io/FileUtils.h
@@ -34,8 +34,9 @@ public:
     static bool readable(const char* path,
                          const char* log = 0);
 
-    static size_t size(const char* path,
-                       const char* log = 0);
+    /// Return the size of the file or -1 on error
+    static int size(const char* path,
+                    const char* log = 0);
 
     /// Make sure the given path is absolute, prepending the current
     /// directory if necessary.
